\name{domainsignatures-package}
\alias{domainsignatures-package}
\alias{domainsignatures}
\docType{package}
\title{
Geneset enrichment based on InterPro domain signatures.
}
\description{
  Compute similarities to pathways for a set of entrezgene identifiers
based on the InterPro domain signature
}
\details{
\tabular{ll}{
Package: \tab domainsignatures\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-07-02\cr
License: \tab LPGL?\cr
}
see help for \code{\link{gseDomain}} for details
}
\author{
Florian Hahne, Tim Beissbarth

Maintainer: Florian Hahne <f.hahne@dkfz.de>
}
\keyword{ package }
